Resolved issue : TimeOutException

+++ app/src/main/java/com/shalatan/entertainmentapp/model/CompleteMovieDetail.kt
@@ -18,7 +18,7 @@ data class CompleteMovieDetail(
     val popularity: Double?,
     val poster_path: String?,
     val release_date: String?,
-    val revenue: Int?,
+    val revenue: Double?,
     val reviews: Reviews?,
     val runtime: Int?,
     val status: String?,
